fix: 修复合并代码是的异常问题
