+++ bcs-app/backend/apps/cluster/serializers.py
@@ -24,7 +24,6 @@ from backend.apps.cluster import constants as cluster_constants
 from backend.apps.cluster.models import ClusterInstallLog, NodeLabel, NodeStatus, NodeUpdateLog
 from backend.components import data as data_api
 from backend.components import paas_cc
-from backend.utils import cc
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.exceptions import ResNotFoundError
